requirements.txt
Django==1.4.3
django-positions==0.4.3
